# openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/users/":{"get":{"summary":"Read Users","operationId":"read_users_users__get","parameters":[{"required":false,"schema":{"title":"Limit","type":"integer","default":10},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read Users Users  Get","type":"array","items":{"$ref":"#/components/schemas/User"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Create User","operationId":"create_user_users__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}":{"get":{"summary":"Read User","operationId":"read_user_users__user_id__get","parameters":[{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}/addresses/":{"post":{"summary":"Create Address","operationId":"create_address_users__user_id__addresses__post","parameters":[{"required":true,"schema":{"title":"User Id","type":"integer"},"name":"user_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Address"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/addresses/":{"get":{"summary":"Read Addresses","operationId":"read_addresses_addresses__get","parameters":[{"required":false,"schema":{"title":"Limit","type":"integer","default":10},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Read Addresses Addresses  Get","type":"array","items":{"$ref":"#/components/schemas/Address"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/addresses/{address_id}":{"get":{"summary":"Read Address","operationId":"read_address_addresses__address_id__get","parameters":[{"required":true,"schema":{"title":"Address Id","type":"integer"},"name":"address_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Address"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update Address","operationId":"update_address_addresses__address_id__put","parameters":[{"required":true,"schema":{"title":"Address Id","type":"integer"},"name":"address_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Address"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Address","operationId":"delete_address_addresses__address_id__delete","parameters":[{"required":true,"schema":{"title":"Address Id","type":"integer"},"name":"address_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/proximity":{"post":{"summary":"Get Proximity","operationId":"get_proximity_proximity_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProximityFind"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Proximity Proximity Post","type":"array","items":{"$ref":"#/components/schemas/Proximity"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Address":{"title":"Address","required":["address","coordinates","id","owner_id","date_created","date_last_updated"],"type":"object","properties":{"address":{"title":"Address","type":"string"},"coordinates":{"title":"Coordinates","type":"string"},"id":{"title":"Id","type":"integer"},"owner_id":{"title":"Owner Id","type":"integer"},"date_created":{"title":"Date Created","type":"string","format":"date-time"},"date_last_updated":{"title":"Date Last Updated","type":"string","format":"date-time"}}},"AddressCreate":{"title":"AddressCreate","required":["address","coordinates"],"type":"object","properties":{"address":{"title":"Address","type":"string"},"coordinates":{"title":"Coordinates","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Proximity":{"title":"Proximity","required":["latitude","longitude","address"],"type":"object","properties":{"latitude":{"title":"Latitude","type":"number"},"longitude":{"title":"Longitude","type":"number"},"address":{"title":"Address","type":"string"}}},"ProximityFind":{"title":"ProximityFind","required":["latitude","longitude","point"],"type":"object","properties":{"latitude":{"title":"Latitude","type":"number"},"longitude":{"title":"Longitude","type":"number"},"point":{"title":"Point","type":"number"}}},"User":{"title":"User","required":["email","id","is_active"],"type":"object","properties":{"email":{"title":"Email","type":"string"},"id":{"title":"Id","type":"integer"},"is_active":{"title":"Is Active","type":"boolean"},"posts":{"title":"Posts","type":"array","items":{"$ref":"#/components/schemas/Address"},"default":[]}}},"UserCreate":{"title":"UserCreate","required":["email","password"],"type":"object","properties":{"email":{"title":"Email","type":"string"},"password":{"title":"Password","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}